requirements.txt
absl-py==2.1.0
array_record==0.5.1
astunparse==1.6.3
biopython==1.84
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.88
click==8.1.7
clu==0.0.12
dm-tree==0.1.8
docstring_parser==0.16
editdistance==0.8.1
etils==1.11.0
flatbuffers==24.3.25
flax==0.10.2
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
jax==0.4.37
jaxlib==0.4.36
keras==3.7.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
ml_collections==1.0.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.0.2
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.10.2
packaging==24.2
promise==2.3
protobuf==3.20.3
psutil==6.1.0
pyglove==0.4.4
Pygments==2.18.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.14.1
sentencepiece==0.2.0
seqio==0.0.19
simplejson==3.19.3
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-text==2.18.0
tensorstore==0.1.69
termcolor==2.5.0
tfds-nightly==4.9.2.dev202308090034
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
